+++ src/main/kotlin/com/yourssu/soongpt/domain/rating/business/dto/RatingResponse.kt
@@ -6,7 +6,7 @@ data class RatingResponse(
     val course: String,
     val professor: String,
     val courseCode: String,
-    val star: Int,
+    val star: Double,
     val point: Double,
 ) {
     companion object {
